+++ conf/dionaea.cfg.cmake
@@ -7,6 +7,11 @@ listen.mode=getifaddrs
 # listen.addresses=127.0.0.1
 # listen.interfaces=eth0,tap0
 
+# Use IPv4 mapped IPv6 addresses
+# It is not recommended to use this feature, try to use nativ IPv4 and IPv6 adresses
+# Valid values: true|false
+# listen.use_ipv4_mapped_ipv6=false
+
 # Country
 # ssl.default.c=GB
 # Common Name/domain name